pyproject.toml
[project]
name = "teaching-vms-subjective-concepts-with-language"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.4.2",
    "dotenv>=0.9.9",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "kaggle>=1.8.3",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "transformers>=4.57.3",
]


[project.optional-dependencies]
cpu = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu128 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
